Move trace_inputlocator out of test folder
